pyproject.toml
[project]
name = "jinchanchan-assistant"
version = "0.1.0"
description = "金铲铲之战 AI 助手 - 支持 Mac PlayCover 和 Windows 模拟器"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jinchanchan Team"}
]
keywords = ["game", "automation", "tft", "ai", "llm"]

dependencies = [
    # 核心依赖
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",

    # 跨平台截图
    "mss>=9.0.0",

    # 配置
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
control = [
    "pyautogui>=0.9.54",
    "pynput>=1.7.6",
]

ocr = [
    "rapidocr-onnxruntime>=1.2.0",
    "opencv-python>=4.8.0",
]

llm = [
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "httpx>=0.26.0",
]

mac = [
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-ApplicationServices>=10.0",
]

windows = [
    "adafruit-circuitpython-adb-shell>=1.0.0",
]

dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-Pillow>=10.0.0",
    "types-PyYAML>=6.0",
]

all = [
    "jinchanchan-assistant[control,ocr,llm,mac,windows]",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "platforms*"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401", "F811"]
# 忽略条件导入块中的未使用导入
"platforms/mac_playcover/adapter.py" = ["F401"]
"core/vision/ocr_engine.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["venv", "tests"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
